moved dinoSearch file
